# Significantly speed up glyph lookups for CJK fonts