@olikraus 2018-04-30
I have started to setup gperf for this (unfortunately gcc-7 does not work with gperf, so I had to go back to gcc-4.8). Results are as expected:
```
Each sample counts as 0.01 seconds.
  %   cumulative   self              self     total           
 time   seconds   seconds    calls  ms/call  ms/call  name    
 47.36      1.88     1.88    84357     0.02     0.02  u8g2_font_get_glyph_data
 36.78      3.34     1.46  5652480     0.00     0.00  u8g_sdl_set_pixel
  2.64      3.45     0.11  1286913     0.00     0.00  u8g2_ll_hvline_vertical_top_lsb
  2.52      3.55     0.10  1794690     0.00     0.00  u8g2_font_decode_len
  2.52      3.65     0.10   706560     0.00     0.00  u8g_sdl_set_8pixel
  2.02      3.73     0.08  2844296     0.00     0.00  u8g2_font_decode_get_unsigned_bits
  1.51      3.79     0.06  2122647     0.00     0.00  u8g2_DrawHVLine
  1.01      3.83     0.04  1732900     0.00     0.00  u8g2_draw_hv_line_4dir
  0.76      3.86     0.03    84357     0.00     0.01  u8g2_font_decode_glyph
  0.50      3.88     0.02  1732900     0.00     0.00  u8g2_draw_l90_r0
  0.50      3.90     0.02   882680     0.00     0.00  u8g2_draw_hv_line_2dir
  0.50      3.92     0.02    11817     0.00     0.00  u8x8_GetStringLineStart
  0.50      3.94     0.02     5520     0.00     0.29  u8g_sdl_set_multiple_8pixel
  0.25      3.95     0.01   253071     0.00     0.00  u8g2_font_decode_get_signed_bits
  0.25      3.96     0.01    84357     0.00     0.00  u8g2_font_setup_decode
  0.25      3.97     0.01       17     0.59     0.59  u8g2_font_get_byte
  0.13      3.97     0.01                             u8g2_DrawLine
  0.00      3.97     0.00  2206965     0.00     0.00  u8g2_add_vector_x
  0.00      3.97     0.00  2206965     0.00     0.00  u8g2_add_vector_y
  0.00      3.97     0.00   653118     0.00     0.00  u8g2_clip_intersection
  0.00      3.97     0.00   264186     0.00     0.00  u8x8_utf8_next
  0.00      3.97     0.00   116806     0.00     0.00  u8g2_is_intersection_decision_tree
  0.00      3.97     0.00    84357     0.00     0.03  u8g2_DrawGlyph
```