@olikraus 2018-05-01
I have updated the code. The lookup table is now part of the new font format.